requirements.txt
Flask==2.3.2
Flask_Cors==3.0.10
Flask_Cors==4.0.0
python_bcrypt==0.3.2
SQLAlchemy==2.0.6
api==0.0.7
asposestorage==1.0.2
python_bcrypt==0.3.2
